+++ pages/services/nft-mint.js
@@ -4,8 +4,8 @@ import SEO from '../../components/SEO'
 import { getIsSsrMobile } from '../../utils/mobile'
 import { xahauNetwork } from '../../utils'
 import NetworkTabs from '../../components/Tabs/NetworkTabs'
-import URITokenMint from '../../components/services/nft-mint/URITokenMint'
-import NFTokenMint from '../../components/services/nft-mint/NFTokenMint'
+import URITokenMint from '../../components/Services/NftMint/URITokenMint'
+import NFTokenMint from '../../components/Services/NftMint/NFTokenMint'
 
 export const getServerSideProps = async (context) => {
   const { query, locale } = context
